+++ OffLogs.Business/Db/Entities/LogTraceEntity.cs
@@ -9,7 +9,7 @@ namespace OffLogs.Business.Db.Entities
     public class LogTraceEntity
     {
         [Id(Name = "Id", Generator = "native")]
-        [Column(Name = "id", SqlType = "INT", NotNull = true)]
+        [Column(Name = "id", SqlType = "bigint", NotNull = true)]
         public virtual long Id { get; set; }
         
         [JsonIgnore]
