+++ OffLogs.Business/Db/Entities/UserEntity.cs
@@ -7,7 +7,7 @@ namespace OffLogs.Business.Db.Entities
     public class UserEntity
     {
         [Id(Name = "Id", Generator = "native")]
-        [Column(Name = "id", SqlType = "INT", NotNull = true)]
+        [Column(Name = "id", SqlType = "bigint", NotNull = true)]
         public virtual long Id { get; set; }
         
         [Property(NotNull = true)]
